[test] moved tests to folder "tests"
