+++ lib/triangle.py
