+++ tests/__init__.py
@@ -1,0 +1,1 @@
+# python.exe -m unittest
