+++ tests/test_square.py
@@ -1,5 +1,5 @@
 import unittest
-import square
+from lib import square
 
 
 class SquareTestCase(unittest.TestCase):
@@ -32,8 +32,9 @@ class SquareTestCase(unittest.TestCase):
         self.assertEqual(res, 1040400)
 
     def test_wrong_area(self):
-        res = square.perimeter("21sd")
-        self.assertEqual(res, 0)
+        def test_wrong_area(self):
+            with self.assertRaises(Exception):
+                square.area("fef")
     def test_zero_perimeter(self):
         res = square.perimeter(0)
         self.assertEqual(res, 0)
@@ -63,5 +64,5 @@ class SquareTestCase(unittest.TestCase):
         self.assertEqual(res, 4080)
 
     def test_wrong_perimeter(self):
-        res = square.perimeter("21sd231")
-        self.assertEqual(res, 0)
+        with self.assertRaises(TypeError):
+            square.perimeter("feffew1")
